requirements.txt
streamlit==1.35.0
crewai==0.55.2
langchain==0.2.12
langchain-community==0.2.10
langchain-huggingface==0.0.3
huggingface-hub==0.23.4
sentence-transformers==3.0.1
chromadb>=0.5.0
groq>=0.9.0
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
pydantic==2.8.2
pytest>=7.0.0